+++ header/minishell_parsing.h
@@ -32,6 +32,7 @@ typedef struct s_parse
 }t_parse;
 
 /* FUNCTIONS */
+int is_redarection(t_lst *node);
 void redarection_prac_fix(t_lst **head);
 void redarection_perfix_lst(t_lst **head);
 void redarection_join_arg(t_lst **node);
